+++ package.json
@@ -23,8 +23,8 @@
   },
   "dependencies": {
     "@file-services/node": "^2.1.1",
-    "@stylable/core": "^3.1.3",
-    "@stylable/language-service": "^3.1.3",
+    "@stylable/core": "^3.1.4",
+    "@stylable/language-service": "^3.1.4",
     "vscode-languageclient": "^6.1.0",
     "vscode-languageserver": "^6.1.0",
     "vscode-languageserver-protocol": "^3.15.2",
