+++ src/lib/vscode-service.ts
@@ -8,6 +8,7 @@ import {
     CompletionParams,
     CompletionItem,
     Definition,
+    Diagnostic,
     DocumentColorParams,
     Hover,
     IConnection,
@@ -23,11 +24,9 @@ import {
 } from 'vscode-languageserver';
 import { URI } from 'vscode-uri';
 
-import { diagnose } from './diagnose';
-
 export class VscodeStylableLanguageService {
     public textDocuments: TextDocuments<TextDocument>;
-    protected languageService: StylableLanguageService;
+    public languageService: StylableLanguageService;
     private connection: IConnection;
 
     constructor(connection: IConnection, docs: TextDocuments<TextDocument>, fs: IFileSystem, stylable: Stylable) {
@@ -86,23 +85,51 @@ export class VscodeStylableLanguageService {
             : [];
     }
 
-    public createDiagnosticsHandler() {
-        const diagnoseConfig = {
-            connection: this.connection,
-            cssService: this.languageService.cssService,
-            docsDispatcher: this.textDocuments,
-            stylable: this.languageService.getStylable()
-        };
-        return () => diagnose(diagnoseConfig);
+    public async diagnoseWithVsCodeConfig() {
+        let res: any;
+        let ignore = false;
+        try {
+            res = await this.connection.workspace.getConfiguration({
+                section: 'stylable'
+            });
+            if (!!res && !!res.diagnostics && !!res.diagnostics.ignore && !!res.diagnostics.ignore.length) {
+                ignore = true;
+            }
+        } catch (e) {
+            /*Client has no workspace/configuration method, ignore silently */
+        }
+
+        const result: Diagnostic[] = [];
+        this.textDocuments.keys().forEach(key => {
+            const doc = this.textDocuments.get(key);
+            if (doc) {
+                if (doc.languageId === 'stylable') {
+                    const fsPath = URI.parse(doc.uri).fsPath;
+                    let diagnostics: Diagnostic[];
+                    if (
+                        ignore &&
+                        (res.diagnostics.ignore as string[]).some(p => {
+                            return fsPath.startsWith(p);
+                        })
+                    ) {
+                        diagnostics = [];
+                    } else {
+                        diagnostics = this.languageService.diagnose(fsPath);
+                        result.push(...diagnostics);
+                    }
+                    this.connection.sendDiagnostics({ uri: doc.uri, diagnostics });
+                }
+            }
+        });
+
+        return result;
     }
 
-    public onDidClose() {
-        return (event: TextDocumentChangeEvent<TextDocument>) => {
-            this.connection.sendDiagnostics({
-                diagnostics: [],
-                uri: event.document.uri
-            });
-        };
+    public onDidClose(event: TextDocumentChangeEvent<TextDocument>) {
+        return this.connection.sendDiagnostics({
+            diagnostics: [],
+            uri: event.document.uri
+        });
     }
 
     private getDocAndPath(uri: string) {
